PSP-1254 DB Schema to align with MOTI Standards (#91)
* PSP-911 Merged DB migrations

* PSP-1254 Apply MOTI DB Schema Standards

* Delete project features
Added ManyToMany properties

* Update project versions
Override Tenant Config
Fix property classifications
Fix property types

* Fix unit tests

+++ backend/api/Areas/Tools/Models/Geocoder/AddressModel.cs
@@ -10,7 +10,7 @@ namespace Pims.Api.Areas.Tools.Models.Geocoder
         public string ProvinceCode { get; set; }
         public double Latitude { get; set; }
         public double Longitude { get; set; }
-        public int Score { get; set; }
+        public long Score { get; set; }
         #endregion
     }
 }
